Add new base history for entire er repo
er started out as a fork of another repository, but I quickly
rewrote so much code of the broken original implementation,
that the entire original repository ceased to have any code
in this fork.

But, github still showed the original parentage of this repository
which made it look like er is "just a fork" instead of an
entire project full of creativity, insight, and
developmental awesomeness.

So, under great pains, I'm truncating the history of er to this
date.  This will render all commit references to the prior
repository useless, for which I am sorry.  You are better off
deleting your old repository and re-checking out this one.

Please forward all complaints to management.